ayos na lahat

+++ database/migrations/2026_03_04_020821_create_abouts_table.php
@@ -11,16 +11,21 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (!Schema::hasTable('abouts')) {
         Schema::create('abouts', function (Blueprint $table) {
             $table->id('about_id');
             $table->string('mission', 255);
             $table->string('vission', 255);
-            $table->foreignId('leadership_id')
-                ->constrained('admin_leaderships', 'leadership_id')
-                ->onDelete('cascade');
+            $table->unsignedBigInteger('leadership_id');
             $table->string('about_desc', 255);
             $table->timestamps();
-        });
+
+            $table->foreign('leadership_id')
+                  ->references('leadership_id')
+                  ->on('admin_leaderships')
+                  ->onDelete('cascade');
+            });
+        }
     }
 
     /**
